add 2084-bit group, use smaller groups in tests
Some tests require that we use the 4096-bit group used in 1Password, but
other tests we can use a smaller group to run tests faster.

So then I noticed some failures trying to use the 2048 bit group. Turns
out we never fully defined it.

A 2048 bit group is acceptable for things like session keys. (Though I
still wouldn't build anything on less than the 3072 bit group). So we
can allow it as an option. It is, however, the minimum, this package
accepts. (As specified in MinGroupSize).