fix: errcheck and gosec linters for srp.go
We were not checking for errors in some locations.

$ gosec ./...
(CWE-703): Errors unhandled.  (Confidence: HIGH, Severity: LOW)
Issues : 5

I'm going to give a shot at refactoring the func signature on a separate
PR as it will be more opinionated than the rest of the changes I'm
proposing here. I think we should be returning an error for newSRP()
and EphemeralPublic().