fix: rm dead code
These helper funcs are meant for testing purposes only. But they are
not used in the SRP repo, nor in other repos consuming the library.
They are simple enought that they can be implemented by the consumers
directly if need be.

I suggest we remove them, as the less code we have to maintain the better.